Test infrastucture for backend
