+++ backend/tests/test_serial_server.py
@@ -1,0 +1,2 @@
+def test():
+    assert 5 == 5
